update to angular rc4

+++ scripts/install-staged.sh
@@ -1,7 +1,8 @@
 #!/usr/bin/env bash
 set -ex
 
-for package in ~/material2/deploy/*
-do
-  npm install ${package}
+for package in ~/material2/deploy/* ; do
+  if [ -f ${package}/package.json ]; then
+    npm install ${package}
+  fi
 done
